+++ ChapeauModel/Product.cs
@@ -9,16 +9,14 @@ namespace ChapeauModel
 {
     public class Product
     {
-        
-
         public int Artikelid { get; private set; }
         public string Naam { get; private set; }
         public int Voorraad { get; private set; }
-        public double Prijs { get; private set; }
+        public decimal Prijs { get; private set; }
         public string Categorie { get; private set; }
         public string Kaart { get; private set; }
 
-        public Product(int artikelid, string naam, int voorraad, double prijs, string categorie, string kaart)
+        public Product(int artikelid, string naam, int voorraad, decimal prijs, string categorie, string kaart)
         {
             Artikelid = artikelid;
             Naam = naam;
@@ -27,7 +25,5 @@ namespace ChapeauModel
             Categorie = categorie;
             Kaart = kaart;
         }
-
-        
     }
 }
